requirements.txt
pdfminer
django
djangorestframework
